pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "docketanalyzer"
version = "0.6.0"
authors = [
    { name = "Nathan Dahlberg" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awscli>=1.38.11",
    "boto3>=1.37.11",
    "elasticsearch==8.12.1",
    "litellm>=1.63.6",
    "pandas>=2.2.3",
    "pathlib>=1.0.1",
    "peewee>=3.17.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "redis>=6.2.0",
    "regex>=2.5.148",
    "requests>=2.32.3",
    "ruff>=0.9.10",
    "simplejson>=3.19.3",
    "tomli>=2.2.1",
    "tomli-w>=1.2.0",
    "tqdm>=4.66.2",
    "twine>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-pygments>=0.0.2",
    "pytest>=8.3.5",
    "pytest-cov>=4.0",
]
all = [
    "datasets>=3.6.0",
    "dill",
    "doclayout_yolo",
    "huggingface-hub",
    "juriscraper>=2.6.55",
    "pymupdf",
    "runpod",
    "scikit-learn>=1.7.0",
    "selenium>=4.18.1",
    "surya-ocr>=0.14.6",
    "transformers[torch]>=4.53.0",
    "uv",
    "uvicorn",
]
ml = [
    "datasets>=3.6.0",
    "scikit-learn>=1.7.0",
    "transformers[torch]>=4.53.0",
]
ocr = [
    "dill",
    "doclayout_yolo",
    "huggingface-hub",
    "pymupdf",
    "runpod",
    "surya-ocr>=0.14.6",
    "uv",
    "uvicorn",
]
pacer = [
    "juriscraper>=2.6.55",
    "selenium>=4.18.1",
]

[project.scripts]
da = "docketanalyzer:cli.cli"

[tool.hatch.build.targets.wheel]
packages = [
    "docketanalyzer",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "*",
    "!docketanalyzer/**",
    "!pyproject.toml",
    "!README.md",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
    "N",
    "SIM",
    "PD",
    "NPY",
    "PTH",
    "RUF",
    "D",
]
ignore = [
    "B008",
    "D100",
    "D104",
    "N801",
    "N806",
]

[tool.ruff.lint.isort]
known-first-party = [
    "docketanalyzer",
]

section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "I001",
    "I002",
    "F403",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
addopts = "-ra -q --cov=docketanalyzer -m 'not cost'"
testpaths = [
    "tests",
]
pythonpath = "."
norecursedirs = "projects"
markers = [
    "cost: tests that incur real costs when run",
    "local: requires credentials only available locally",
]
